refactor(components): reorganize component structure
Reorganizes components into domain-based structure
Normalizes layout and naming inconsistencies
Improves maintainability and discoverability
No behavioral changes
