+++ src/app/globals.css
@@ -8,7 +8,7 @@
 @import '../styles/comments.css';
 @import '../styles/resume.css';
 @import '../styles/mobile.css';
-@import '../components/Layouts/Nav/styles.css';
-@import '../components/ProjectCard/style.css';
-@import '../components/LanguagesSprites/styles.css';
-@import '../components/ContactForm/style.css';
+@import '../components/layout/nav/styles.css';
+@import '../components/projects/style.css';
+@import '../components/shared/languages/styles.css';
+@import '../components/contact/style.css';
