+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import {BaseWrapper} from '@/components/layout/BaseWrapper'
-import MainSection from '@/components/Sections/MainSection/MainSection'
-import AboutSection from '@/components/Sections/AboutSection'
-import {ContactForm} from '@/components/ContactForm'
+import MainSection from '@/components/sections/main/MainSection'
+import AboutSection from '@/components/sections/AboutSection'
+import {ContactForm} from '@/components/contact'
 
 export default function HomePage() {
   return (
