--- src/components/ContactForm/index.tsx
@@ -1,1 +1,0 @@
-export {default as ContactForm} from '@/components/ContactForm/NewContactForm'
