--- src/components/FormSucess/Index.tsx
@@ -1,1 +1,0 @@
-export {default as FormSuccess} from '@/components/FormSucess/FormSucess'
