--- src/components/Layouts/Nav/index.tsx
@@ -1,1 +1,0 @@
-export {Nav} from '@/components/Layouts/Nav/Nav'
