--- src/components/ProjectCard/index.tsx
@@ -1,1 +1,0 @@
-export {default as ProjectCard} from '@/components/ProjectCard/ProjectCard'
