--- src/components/Sections/index.tsx
@@ -1,3 +1,0 @@
-export {default as MainSection} from '@/components/Sections/MainSection/MainSection'
-export {default as AboutSection} from '@/components/Sections/AboutSection'
-export {default as ProjectSection} from '@/components/Sections/ProjectSection'
