+++ src/components/contact/index.tsx
@@ -1,0 +1,1 @@
+export {default as ContactForm} from '@/components/contact/NewContactForm'
