+++ src/components/contact/success/index.tsx
@@ -1,0 +1,1 @@
+export {default as FormSuccess} from '@/components/contact/success/FormSuccess'
