+++ src/components/layout/nav/index.tsx
@@ -1,0 +1,1 @@
+export {Nav} from '@/components/layout/nav/Nav'
