+++ src/components/projects/index.tsx
@@ -1,0 +1,1 @@
+export {default as ProjectCard} from '@/components/projects/ProjectCard'
