+++ src/components/sections/AnimatedSection.tsx
