+++ src/components/sections/index.tsx
@@ -1,0 +1,3 @@
+export {default as MainSection} from '@/components/sections/main/MainSection'
+export {default as AboutSection} from '@/components/sections/AboutSection'
+export {default as ProjectSection} from '@/components/sections/ProjectSection'
